backend/requirements.txt
asgiref==3.8.1
Django==4.2.15
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
requests==2.32.3
setuptools==74.0.0
sqlparse==0.5.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2

